Removes some code duplication.
